Add FormTitle as a variant of BaseTitle for forms
