+++ src/components/Form/index.ts
@@ -1,1 +1,2 @@
+export { default as FormTitle } from "./FormTitle";
 export { default as InputField } from "./InputField";
